protect agains delete and update of resources that are part of a set that is collating. bump rev to release candidate 1

+++ juxta-ws/src/main/java/org/juxtasoftware/resource/ComparisonSetResource.java
@@ -210,9 +210,14 @@ public class ComparisonSetResource extends BaseResource {
     }
    
     @Delete
-    public void deleteComparisonSet() {
+    public Representation deleteComparisonSet() {
         LOG.info("Delete set "+this.set.getId());
+        if ( this.set.getStatus().equals(ComparisonSet.Status.COLLATING)) {
+            setStatus(Status.CLIENT_ERROR_CONFLICT);
+            return toTextRepresentation("Cannot delete set; collation is in progress");
+        }
         this.comparionSetDao.delete(set);
+        return toTextRepresentation("ok");
     }
 
     private static class SourceExclusion implements ExclusionStrategy {
